renamed common -> _common and started app_creator
