+++ fastgenomics/_common.py
