+++ tests/test_common.py
@@ -1,49 +1,49 @@
 import pytest
 import pathlib
 
-from fastgenomics import common
+from fastgenomics import _common
 
 
 def test_paths_are_initialized(local):
-    common.get_paths()
+    _common.get_paths()
 
 
 def test_custom_init_paths(app_dir, data_root):
-    common.set_paths(app_dir, data_root)
-    common.get_paths()
+    _common.set_paths(app_dir, data_root)
+    _common.get_paths()
 
 
 def test_paths_from_env(fg_env):
-    common.get_paths()
+    _common.get_paths()
 
 
 def test_cannot_init_nonexisting_paths():
     with pytest.raises(RuntimeError):
-        common.set_paths("i_don't_exist", "me_neither")
+        _common.set_paths("i_don't_exist", "me_neither")
 
 
 def test_custom_init_path_within_docker(fake_docker, app_dir, data_root):
     with pytest.warns(None):
-        common.set_paths(app_dir, data_root)
-        common.get_paths()
+        _common.set_paths(app_dir, data_root)
+        _common.get_paths()
 
 
 def test_get_app_manifest(local):
-    common.get_app_manifest()
+    _common.get_app_manifest()
 
 
 def test_assert_manifest_is_valid(local):
-    manifest = common.get_app_manifest()
-    common.assert_manifest_is_valid({'FASTGenomicsApplication': manifest})
+    manifest = _common.get_app_manifest()
+    _common.assert_manifest_is_valid({'FASTGenomicsApplication': manifest})
 
 
 def test_can_get_parameters(local):
-    parameters = common.get_parameters()
+    parameters = _common.get_parameters()
     assert len(parameters) > 0
 
 
 def test_parameters(local):
-    parameters = common.get_parameters()
+    parameters = _common.get_parameters()
 
     assert "StrValue" in parameters
     assert parameters["StrValue"] == "hello from parameters.json"
@@ -65,60 +65,60 @@ def test_parameters(local):
 
 
 def test_can_get_specific_parameter(local):
-    assert common.get_parameter("IntValue") == 150
+    assert _common.get_parameter("IntValue") == 150
 
 
 def test_can_have_different_type(local, monkeypatch):
     # patch custom parameter load function
-    monkeypatch.setattr("fastgenomics.common.load_runtime_parameters", lambda: {"StrValue": 1})
+    monkeypatch.setattr("fastgenomics._common.load_runtime_parameters", lambda: {"StrValue": 1})
 
     # get parameters and compare parameters of different types
     with pytest.warns(None):
-        parameters = common.get_parameters()
+        parameters = _common.get_parameters()
         assert 1 == parameters["StrValue"]
 
 
 def test_load_input_file_mapping(local):
-    input_file_mapping = common.load_input_file_mapping()
+    input_file_mapping = _common.load_input_file_mapping()
     assert "some_input" in input_file_mapping
 
 
 def test_input_file_mapping_to_paths(local):
-    ifm_dict = common.load_input_file_mapping()
-    input_file_mapping = common.str_to_path_file_mapping(ifm_dict)
+    ifm_dict = _common.load_input_file_mapping()
+    input_file_mapping = _common.str_to_path_file_mapping(ifm_dict)
     assert isinstance(input_file_mapping['some_input'], pathlib.Path)
     assert input_file_mapping['some_input'].exists()
 
 
 def test_check_input_file_mapping(local):
-    ifm_dict = common.load_input_file_mapping()
-    input_file_mapping = common.str_to_path_file_mapping(ifm_dict)
+    ifm_dict = _common.load_input_file_mapping()
+    input_file_mapping = _common.str_to_path_file_mapping(ifm_dict)
 
     # test everything is ok
-    common.check_input_file_mapping(input_file_mapping)
+    _common.check_input_file_mapping(input_file_mapping)
 
     # test if additional keys trigger warning
     input_file_mapping['unused_key'] = pathlib.Path(".")
     with pytest.warns(None):
-        common.check_input_file_mapping(input_file_mapping)
+        _common.check_input_file_mapping(input_file_mapping)
 
     # test raises KeyError on missing entry
     with pytest.raises(KeyError):
-        common.check_input_file_mapping({})
+        _common.check_input_file_mapping({})
 
     # test raises FileNotFoundError on wrong item
     with pytest.raises(FileNotFoundError):
-        common.check_input_file_mapping({"some_input": pathlib.Path("i_don't_exist")})
+        _common.check_input_file_mapping({"some_input": pathlib.Path("i_don't_exist")})
 
 
 def test_load_input_file_mapping_from_env(local, monkeypatch):
     monkeypatch.setenv('INPUT_FILE_MAPPING', '{"some_key": "some_value"}')
-    input_file_mapping = common.load_input_file_mapping()
+    input_file_mapping = _common.load_input_file_mapping()
     assert "some_input" not in input_file_mapping, "input_file_mapping from file used instead if env!"
     assert "some_key" in input_file_mapping
 
 
 def test_get_input_file_mapping(local):
-    input_file_mapping = common.get_input_file_mapping()
+    input_file_mapping = _common.get_input_file_mapping()
     assert "some_input" in input_file_mapping
     assert input_file_mapping['some_input'].exists()
